Refactoring fix namespace with typo.
