+++ core/booking/tests/Integration/ReservationRepositoryExpiredReadSideTest.php
@@ -5,7 +5,7 @@ namespace Booking\Tests\Integration;
 
 use App\Factory\ReservationFactory;
 use App\Factory\ReservationSaleDetailFactory;
-use Booking\Adapter\Persistance\ReservationRepository;
+use Booking\Adapter\Persistence\ReservationRepository;
 use Booking\Application\Domain\Model\Reservation;
 use DateTimeImmutable;
 use DateTimeZone;
@@ -14,7 +14,7 @@ use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Zenstruck\Foundry\Test\Factories;
 use Zenstruck\Foundry\Test\ResetDatabase;
 
-/** @covers \Booking\Adapter\Persistance\ReservationRepository */
+/** @covers \Booking\Adapter\Persistence\ReservationRepository */
 class ReservationRepositoryExpiredReadSideTest extends KernelTestCase
 {
     /**
