+++ phpstan-baseline.neon
@@ -51,69 +51,69 @@ parameters:
 			path: core/booking/src/Adapter/MailDriven/BookingMailer.php
 
 		-
-			message: "#^Class Booking\\\\Adapter\\\\Persistance\\\\BookRepository extends generic class Doctrine\\\\Bundle\\\\DoctrineBundle\\\\Repository\\\\ServiceEntityRepository but does not specify its types\\: T$#"
+			message: "#^Class Booking\\\\Adapter\\\\Persistence\\\\BookRepository extends generic class Doctrine\\\\Bundle\\\\DoctrineBundle\\\\Repository\\\\ServiceEntityRepository but does not specify its types\\: T$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/BookRepository.php
+			path: core/booking/src/Adapter/Persistence/BookRepository.php
 
 		-
 			message: "#^Construct empty\\(\\) is not allowed\\. Use more strict comparison\\.$#"
 			count: 2
-			path: core/booking/src/Adapter/Persistance/DoctrineType/ReservationStatusType.php
+			path: core/booking/src/Adapter/Persistence/DoctrineType/ReservationStatusType.php
 
 		-
 			message: "#^Parameter \\#1 \\$name of static method Booking\\\\Application\\\\Domain\\\\Model\\\\ReservationStatus\\:\\:fromName\\(\\) expects string, mixed given\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/DoctrineType/ReservationStatusType.php
+			path: core/booking/src/Adapter/Persistence/DoctrineType/ReservationStatusType.php
 
 		-
-			message: "#^Method Booking\\\\Adapter\\\\Persistance\\\\ExportDataRetriever\\:\\:getAllCustomerForNewsletter\\(\\) return type has no value type specified in iterable type array\\.$#"
+			message: "#^Method Booking\\\\Adapter\\\\Persistence\\\\ExportDataRetriever\\:\\:getAllCustomerForNewsletter\\(\\) return type has no value type specified in iterable type array\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ExportDataRetriever.php
+			path: core/booking/src/Adapter/Persistence/ExportDataRetriever.php
 
 		-
 			message: "#^Argument of an invalid type mixed supplied for foreach, only iterables are supported\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
 			message: "#^Cannot call method getConfirmationStatus\\(\\) on Booking\\\\Application\\\\Domain\\\\Model\\\\ReservationSaleDetail\\|null\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
 			message: "#^Cannot call method isExpired\\(\\) on Booking\\\\Application\\\\Domain\\\\Model\\\\ConfirmationStatus\\\\ConfirmationStatus\\|null\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
 			message: "#^Casting to int something that's already int\\<0, max\\>\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
-			message: "#^Class Booking\\\\Adapter\\\\Persistance\\\\ReservationRepository extends generic class Doctrine\\\\Bundle\\\\DoctrineBundle\\\\Repository\\\\ServiceEntityRepository but does not specify its types\\: T$#"
+			message: "#^Class Booking\\\\Adapter\\\\Persistence\\\\ReservationRepository extends generic class Doctrine\\\\Bundle\\\\DoctrineBundle\\\\Repository\\\\ServiceEntityRepository but does not specify its types\\: T$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
-			message: "#^Method Booking\\\\Adapter\\\\Persistance\\\\ReservationRepository\\:\\:findAllConfirmedOrderByOldest\\(\\) return type has no value type specified in iterable type array\\.$#"
+			message: "#^Method Booking\\\\Adapter\\\\Persistence\\\\ReservationRepository\\:\\:findAllConfirmedOrderByOldest\\(\\) return type has no value type specified in iterable type array\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
-			message: "#^Method Booking\\\\Adapter\\\\Persistance\\\\ReservationRepository\\:\\:findAllConfirmedOrderByOldest\\(\\) should return array but returns mixed\\.$#"
+			message: "#^Method Booking\\\\Adapter\\\\Persistence\\\\ReservationRepository\\:\\:findAllConfirmedOrderByOldest\\(\\) should return array but returns mixed\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
-			message: "#^Method Booking\\\\Adapter\\\\Persistance\\\\ReservationRepository\\:\\:findAllNewArrivalOrderByNewest\\(\\) return type has no value type specified in iterable type array\\.$#"
+			message: "#^Method Booking\\\\Adapter\\\\Persistence\\\\ReservationRepository\\:\\:findAllNewArrivalOrderByNewest\\(\\) return type has no value type specified in iterable type array\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
-			message: "#^Method Booking\\\\Adapter\\\\Persistance\\\\ReservationRepository\\:\\:findAllNewArrivalOrderByNewest\\(\\) should return array but returns mixed\\.$#"
+			message: "#^Method Booking\\\\Adapter\\\\Persistence\\\\ReservationRepository\\:\\:findAllNewArrivalOrderByNewest\\(\\) should return array but returns mixed\\.$#"
 			count: 1
-			path: core/booking/src/Adapter/Persistance/ReservationRepository.php
+			path: core/booking/src/Adapter/Persistence/ReservationRepository.php
 
 		-
 			message: "#^Call to an undefined method Booking\\\\Application\\\\Domain\\\\Model\\\\ReservationRepositoryInterface\\:\\:findWithQueryBuilderAllConfirmedOrderByOldest\\(\\)\\.$#"
@@ -471,7 +471,7 @@ parameters:
 			path: core/booking/tests/Integration/ExportDataRetrieverTest.php
 
 		-
-			message: "#^Cannot call method getAllCustomerForNewsletter\\(\\) on Booking\\\\Adapter\\\\Persistance\\\\ExportDataRetriever\\|null\\.$#"
+			message: "#^Cannot call method getAllCustomerForNewsletter\\(\\) on Booking\\\\Adapter\\\\Persistence\\\\ExportDataRetriever\\|null\\.$#"
 			count: 2
 			path: core/booking/tests/Integration/ExportDataRetrieverTest.php
